backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
google-cloud-aiplatform>=1.70.0
firebase-admin>=6.5.0
requests>=2.32.0
extract-msg>=0.48.0
openpyxl>=3.1.5
pandas>=2.2.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pillow>=10.4.0
pydantic>=2.9.0
